[5.x] Remove Antlers Regex Parser (#9442)
