+++ config/antlers.php
@@ -4,18 +4,6 @@ return [
 
     /*
     |--------------------------------------------------------------------------
-    | Version
-    |--------------------------------------------------------------------------
-    |
-    | The desired Antlers language version to utilize. Supported values are
-    | "runtime" for the modern parser, or "regex" for the legacy parser.
-    |
-    */
-
-    'version' => 'runtime',
-
-    /*
-    |--------------------------------------------------------------------------
     | Guarded Variables
     |--------------------------------------------------------------------------
     |
